requirements.txt
scikit-learn
sastrawi
numpy
pandas
fastapi
uvicorn
python-dotenv
groq